Rework NaturalSort to fix version comparison issues
It now accepts unbalanced versions numbers, for example:

  - 1.0 == 1.0.0.0
  - 1.0 < 1.0.1
  - 1.0.2 > 1.0

The tilde operator was matching invalid version numbers, because we
only checked that the version number started with the loose
requirement, not that it matched the version number segment. For
example "~> 0.1.0" matched "0.10.0" which was invalid.

Added some tests from @straight-shoota for Helpers::Versions and
some more tests for the lower Helpers::NaturalSort module.