Add --local to not fetch remote repositories
Uses the local cahe as is. Speeds up installing shards when we know
the cache is up to date.

It will fail when the repository wasn't cloned into the cache, and
it won't find new releases until the cache is updated by running the
install or update commands without --local.